working on better API to remove Copy trait req
