+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "zarray"
-version = "1.2.0"
+version = "2.0.0"
 edition = "2021"
 authors = ["Christopher Collin Hall"]
 description = """Cache-optimized 2D and 3D arrays using Morton order (aka
@@ -16,7 +16,8 @@ readme = "README.md"
 homepage = "https://github.com/DrPlantabyte/ZArray-rs"
 
 [dependencies]
-# no required dependencies
+array-init = "2"
+#optional deps
 serde = { version = "1.0", optional = true, features = ["derive"] }
 
 [dev-dependencies]
